+++ src/MineJason/ChatComponent.cs
@@ -24,6 +24,7 @@ public abstract class ChatComponent(string? type) : IEquatable<ChatComponent>
     /// <summary>
     /// Gets the serializer options that is required for conforming the Minecraft: Java Edition standards.
     /// </summary>
+    [Obsolete("Deserialize and serialize directly.")]
     public static readonly JsonSerializerOptions SerializerOptions = new()
     {
         PropertyNamingPolicy = JsonNamingPolicy.CamelCase,